+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/webview/in_app_webview/InAppWebView.kt
@@ -872,6 +872,15 @@ class InAppWebView : InputAwareWebView, InAppWebViewInterface {
   }
 
   fun prepareAndAddUserScripts() {
+    if (customSettings.disableAutocorrection == true) {
+      userContentController.addPluginScript(
+        PluginScriptsUtil.DISABLE_AUTOCORRECTION_JS_PLUGIN_SCRIPT(
+          customSettings.pluginScriptsOriginAllowList,
+          customSettings.pluginScriptsForMainFrameOnly == true
+        )
+      )
+    }
+
     if (javaScriptBridgeEnabled) {
       userContentController.addPluginScript(
         PromisePolyfillJS.PROMISE_POLYFILL_JS_PLUGIN_SCRIPT(