+++ flutter_inappwebview_forge_linux/linux/in_app_webview/in_app_webview_settings.cc
@@ -50,6 +50,7 @@ InAppWebViewSettings::InAppWebViewSettings(FlValue* map) : InAppWebViewSettings(
   supportZoom = get_fl_map_value(map, "supportZoom", supportZoom);
   isInspectable = get_fl_map_value(map, "isInspectable", isInspectable);
   disableContextMenu = get_fl_map_value(map, "disableContextMenu", disableContextMenu);
+  disableAutocorrection = get_fl_map_value(map, "disableAutocorrection", disableAutocorrection);
 
   // === JavaScript bridge settings ===
   if (fl_map_contains_not_null(map, "javaScriptHandlersOriginAllowList")) {
@@ -516,6 +517,7 @@ FlValue* InAppWebViewSettings::toFlValue() const {
       {"supportZoom", make_fl_value(supportZoom)},
       {"isInspectable", make_fl_value(isInspectable)},
       {"disableContextMenu", make_fl_value(disableContextMenu)},
+      {"disableAutocorrection", make_fl_value(disableAutocorrection)},
 
       // === JavaScript bridge settings ===
       {"javaScriptBridgeEnabled", make_fl_value(javaScriptBridgeEnabled)},