+++ flutter_inappwebview_forge_linux/linux/in_app_webview/in_app_webview_settings.h
@@ -44,6 +44,7 @@ class InAppWebViewSettings {
   bool supportZoom = true;
   bool isInspectable = true;
   bool disableContextMenu = false;
+  bool disableAutocorrection = false;
 
   // === JavaScript bridge settings ===
   std::optional<std::vector<std::string>> javaScriptHandlersOriginAllowList =