+++ flutter_inappwebview_forge_macos/macos/flutter_inappwebview_forge_macos/Sources/flutter_inappwebview_forge_macos/InAppWebView/InAppWebViewSettings.swift
@@ -53,6 +53,7 @@ public class InAppWebViewSettings: ISettings<InAppWebView> {
     var allowingReadAccessTo: String? = nil
     var underPageBackgroundColor: String?
     var isTextInteractionEnabled = true
+    var disableAutocorrection = false
     var isSiteSpecificQuirksModeEnabled = true
     var upgradeKnownHostsToHTTPS = true
     var isElementFullscreenEnabled = true